make sure onMount effect runs last
