+++ packages/tldraw/src/lib/Tldraw.tsx
@@ -1,27 +1,31 @@
 import {
 	Canvas,
+	Editor,
 	ErrorScreen,
 	LoadingScreen,
 	RecursivePartial,
+	TLOnMountHandler,
 	TldrawEditor,
 	TldrawEditorProps,
+	assert,
+	useEditor,
 } from '@tldraw/editor'
-import { useMemo } from 'react'
+import { useCallback, useDebugValue, useLayoutEffect, useMemo, useRef } from 'react'
 import { TldrawHandles } from './canvas/TldrawHandles'
 import { TldrawHoveredShapeIndicator } from './canvas/TldrawHoveredShapeIndicator'
 import { TldrawScribble } from './canvas/TldrawScribble'
 import { TldrawSelectionBackground } from './canvas/TldrawSelectionBackground'
 import { TldrawSelectionForeground } from './canvas/TldrawSelectionForeground'
+import {
+	TLExternalContentProps,
+	registerDefaultExternalContentHandlers,
+} from './defaultExternalContentHandlers'
 import { defaultShapeTools } from './defaultShapeTools'
 import { defaultShapeUtils } from './defaultShapeUtils'
+import { registerDefaultSideEffects } from './defaultSideEffects'
 import { defaultTools } from './defaultTools'
 import { TldrawUi, TldrawUiProps } from './ui/TldrawUi'
 import { ContextMenu } from './ui/components/ContextMenu'
-import {
-	TLExternalContentProps,
-	useRegisterExternalContentHandlers,
-} from './useRegisterExternalContentHandlers'
-import { useSideEffects } from './useSideEffects'
 import { TLEditorAssetUrls, useDefaultEditorAssetsWithOverrides } from './utils/assetUrls'
 import { usePreloadAssets } from './utils/usePreloadAssets'
 
@@ -42,6 +46,7 @@ export function Tldraw(
 		maxAssetSize,
 		acceptedImageMimeTypes,
 		acceptedVideoMimeTypes,
+		onMount,
 		...rest
 	} = props
 
@@ -94,6 +99,7 @@ export function Tldraw(
 					maxAssetSize={maxAssetSize}
 					acceptedImageMimeTypes={acceptedImageMimeTypes}
 					acceptedVideoMimeTypes={acceptedVideoMimeTypes}
+					onMount={onMount}
 				/>
 			</TldrawUi>
 		</TldrawEditor>
@@ -106,14 +112,53 @@ function InsideOfEditorContext({
 	maxAssetSize = 10 * 1024 * 1024, // 10mb
 	acceptedImageMimeTypes = ['image/jpeg', 'image/png', 'image/gif', 'image/svg+xml'],
 	acceptedVideoMimeTypes = ['video/mp4', 'video/quicktime'],
-}: Partial<TLExternalContentProps>) {
-	useRegisterExternalContentHandlers({
-		maxImageDimension,
-		maxAssetSize,
-		acceptedImageMimeTypes,
-		acceptedVideoMimeTypes,
+	onMount,
+}: Partial<TLExternalContentProps & { onMount: TLOnMountHandler }>) {
+	const editor = useEditor()
+
+	const onMountEvent = useEvent((editor: Editor) => {
+		const unsubs: (void | (() => void) | undefined)[] = []
+
+		unsubs.push(registerDefaultSideEffects(editor))
+
+		// for content handling, first we register the default handlers...
+		registerDefaultExternalContentHandlers(editor, {
+			maxImageDimension,
+			maxAssetSize,
+			acceptedImageMimeTypes,
+			acceptedVideoMimeTypes,
+		})
+
+		// ...then we run the onMount prop, which may override the above
+		unsubs.push(onMount?.(editor))
+
+		return () => {
+			unsubs.forEach((fn) => fn?.())
+		}
 	})
-	useSideEffects()
+
+	useLayoutEffect(() => {
+		if (editor) return onMountEvent?.(editor)
+	}, [editor, onMountEvent])
 
 	return null
+}
+
+// duped from tldraw editor
+function useEvent<Args extends Array<unknown>, Result>(
+	handler: (...args: Args) => Result
+): (...args: Args) => Result {
+	const handlerRef = useRef<(...args: Args) => Result>()
+
+	useLayoutEffect(() => {
+		handlerRef.current = handler
+	})
+
+	useDebugValue(handler)
+
+	return useCallback((...args: Args) => {
+		const fn = handlerRef.current
+		assert(fn, 'fn does not exist')
+		return fn(...args)
+	}, [])
 }
