+++ packages/tldraw/src/lib/defaultSideEffects.ts
@@ -1,0 +1,34 @@
+import { Editor } from '@tldraw/editor'
+
+export function registerDefaultSideEffects(editor: Editor) {
+	return editor.sideEffects.registerAfterChangeHandler('instance_page_state', (prev, next) => {
+		if (prev.croppingShapeId !== next.croppingShapeId) {
+			const isInCroppingState = editor.isInAny(
+				'select.crop',
+				'select.pointing_crop_handle',
+				'select.cropping'
+			)
+			if (!prev.croppingShapeId && next.croppingShapeId) {
+				if (!isInCroppingState) {
+					editor.setCurrentTool('select.crop.idle')
+				}
+			} else if (prev.croppingShapeId && !next.croppingShapeId) {
+				if (isInCroppingState) {
+					editor.setCurrentTool('select.idle')
+				}
+			}
+		}
+
+		if (prev.editingShapeId !== next.editingShapeId) {
+			if (!prev.editingShapeId && next.editingShapeId) {
+				if (!editor.isIn('select.editing_shape')) {
+					editor.setCurrentTool('select.editing_shape')
+				}
+			} else if (prev.editingShapeId && !next.editingShapeId) {
+				if (editor.isIn('select.editing_shape')) {
+					editor.setCurrentTool('select.idle')
+				}
+			}
+		}
+	})
+}
